requirements.txt
django>=4.2,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
django-cors-headers>=4.0
django-mptt>=0.14
psycopg2-binary>=2.9
Pillow==10.1.0
python-barcode==0.15.1
qrcode==7.4.2
celery>=5.3
redis>=5.0
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
